params and some tests

+++ src/index.ts
@@ -2,10 +2,11 @@ import path from 'path'
 import { spawn, ChildProcess } from 'child_process'
 import cliProgress from 'cli-progress'
 import * as dotenv from 'dotenv'
-import fs from 'fs'
+import fs, { stat } from 'fs'
 import chalk from 'chalk'
-import SysMetrics from './SysMetrics'
 import pidusage from 'pidusage';
+import { parseArgs } from "node:util";
+
 dotenv.config()
 
 type FrontierTesterProps = {
@@ -14,7 +15,9 @@ type FrontierTesterProps = {
     numTuples: number
     warmUpTuples: number
     nWorkers?: number
-    queryType: string
+    queryType: string,
+    folder: string,
+    keepalive: boolean,
 }
 class FrontierTester {
     private readonly FRONTIER_PATH = "/home/toscan/dev/bd/frontier"
@@ -23,13 +26,15 @@ class FrontierTester {
     private readonly QUERY_JAR = path.join(this.EXAMPLE_PATH, "dist/acita_demo_2015.jar")
     private readonly LOG_DIR = "./logs"
     private readonly RESULTS_DIR = "./results"
+    private readonly MONITORING_FILE
+    private readonly PIDSMAP
+    private readonly EVENTS_FILE
+    private readonly SINK_FILE
     
     private readonly N_WORKERS
     private readonly N_TUPLES
     private readonly KILL_EACH
     
-    private sysMetrics = new SysMetrics(path.join(this.RESULTS_DIR, 'monitoring.csv'), 100)
-
     private itsKillingTime = false
     private monitoring: ChildProcess | null = null
     private master: ChildProcess | null = null
@@ -37,8 +42,16 @@ class FrontierTester {
     private sinkData: string[] = []
     private readonly props: { [key: string]: string | number }
     private starting = true;
+    private pidNames: { [id: number]: string } = {}
+    private readonly keepAlive: boolean
 
     constructor(props: FrontierTesterProps) {
+        fs.mkdirSync(path.join(this.RESULTS_DIR, props.folder))
+        this.MONITORING_FILE = path.join(this.RESULTS_DIR, props.folder, 'monitoring.csv')
+        this.PIDSMAP = path.join(this.RESULTS_DIR, props.folder, 'pidMap.csv')
+        this.EVENTS_FILE = path.join(this.RESULTS_DIR, props.folder, 'events.csv')
+        this.SINK_FILE = path.join(this.RESULTS_DIR, props.folder, 'sink.csv')
+
         this.props = {
             "replicationFactor": props.replicationFactor,
             "chainLength": props.chainLength,
@@ -49,6 +62,7 @@ class FrontierTester {
         this.N_WORKERS = props.nWorkers ?? props.replicationFactor * props.chainLength
         this.N_TUPLES = props.numTuples
         this.KILL_EACH = Math.ceil(this.N_TUPLES / (this.N_WORKERS)) + 1
+        this.keepAlive = props.keepalive
     }
 
     private progressBar = new cliProgress.SingleBar({}, cliProgress.Presets.shades_classic)
@@ -80,8 +94,42 @@ class FrontierTester {
         }
     }
 
+    private cleaup() {
+        if (fs.existsSync(this.MONITORING_FILE))
+            fs.unlinkSync(this.MONITORING_FILE)
+        if (fs.existsSync(this.EVENTS_FILE))
+            fs.unlinkSync(this.EVENTS_FILE)
+    }
+
+    private sysMonitoring(pids: number[]) {
+        const compute = async () => {
+            const stats = await pidusage(pids)
+            const keys = Object.keys(stats);
+            const firstTimestamp = stats[keys[0]]?.timestamp ?? null
+            if (firstTimestamp === null) {
+                return false
+            }
+            const line = keys.map(pid => ({ s: stats[pid], pid }))
+                            .map(({ s, pid }) => `${pid},${s?.cpu ?? 0},${s?.memory ?? 0}`)
+                            .join("\t")
+            // print all cpu
+            const cc = Object.keys(stats).map(pid => stats[pid]?.cpu ?? -69)
+
+            fs.appendFileSync(this.MONITORING_FILE, `${firstTimestamp}\t${line}\n`)
+            return true
+        }
+        const interval = async (time: number) => {
+            setTimeout(async () => {
+                if (await compute())
+                    interval(time);
+            }, time)
+        }
+        interval(1000)
+    }
+
     public async run() {
-        this.sysMetrics.start();
+        this.cleaup()
+        
         this.master = this.startMaster()
         await this.sleep(3)
         for (let i = 0; i < this.N_WORKERS + 2; i++) {
@@ -91,7 +139,9 @@ class FrontierTester {
             await this.sleep(1)
         }
         // this.monitoring = this.startMonitoring();
-        await this.sleep(2)
+        await this.sleep(1)
+        this.sysMonitoring(this.workers.filter(w => w.pid !== undefined).map(w => w.pid as number))
+        await this.sleep(1)
 
         console.log("Deploying query ... ");
         this.deployQuery();
@@ -101,12 +151,10 @@ class FrontierTester {
     private async done() {
         await this.sleep(1);
 
-        this.sysMetrics.stop();
         this.monitoring?.kill()
         this.master?.kill()
         this.workers.forEach(w => w.kill())
-        const sinkDataPath = path.join(this.RESULTS_DIR, `sink.csv`)
-        fs.writeFileSync(sinkDataPath, this.sinkData.join("\n"))
+        fs.writeFileSync(this.SINK_FILE, this.sinkData.join("\n"))
         console.log("Done")
     }
 
@@ -151,15 +199,6 @@ class FrontierTester {
         return p
     }
 
-    private startMonitoring() {
-        const filename = 'results/cpu.csv'
-        if (fs.existsSync(filename)) fs.unlinkSync(filename);
-        const cmd = `/home/toscan/bin/dool -T -c -C 0,1,2,3,4,5,6,7,8,9,10,11,total -m --output ${filename} 1`
-        console.log( chalk.magenta( 'STARTING MONITORING' ))
-        const p = spawn (cmd, { shell: '/bin/zsh' })
-        return p;
-    }
-
     private deployQuery() {
         this.master?.stdin?.write('7\n')
     }
@@ -174,11 +213,13 @@ class FrontierTester {
                 // this.printColored(line, port)
                 // return
                 const data = line.split(",")
+                this.pidNames[p.pid!] = data[1]
                 if (this.itsKillingTime && data[1] === "PROCESSOR") {
-                    p.kill()
-
-                    console.log("\nKilled worker\n")
-                    this.log('KILLED')
+                    if (!this.keepAlive) {
+                        p.kill()
+                        console.log("\nKilled worker\n")
+                        this.log('KILLED')
+                    }
                     // console.log(line)
                     this.itsKillingTime = false
                 } else if (data[1] === "SINK") {
@@ -192,6 +233,7 @@ class FrontierTester {
                         this.progressBar.stop()
                         this.log('DONE')
                         this.done()
+                        fs.writeFileSync(this.PIDSMAP, Object.keys(this.pidNames).map(k => `${k},${this.pidNames[+k]}`).join("\n"))
                     }
                 }
             }
@@ -199,19 +241,37 @@ class FrontierTester {
     }
 
     private log(name: string) {
-        fs.appendFileSync(
-            path.join(this.RESULTS_DIR, 'events.csv'), 
-            `${new Date().getTime()},${name}\n`
-        )
+        fs.appendFileSync(this.EVENTS_FILE, `${new Date().getTime()},${name}\n`)
     }
 }
 
+const params = parseArgs({
+    options: {
+        folder:         { type: "string", short: "f" },
+        query:          { type: "string", short: "q", default: "chain" },
+        replication:    { type: "string", short: "r", default: "3" },
+        length:         { type: "string", short: "l", default: "1" },
+        tuples:         { type: "string", short: "t", default: "20000" },
+        warmup:         { type: "string", short: "w", default: "0" },
+        keepalive:      { type: "boolean", short: "k" },
+    },
+});
+
+if (!params.values.folder) {
+    console.log("Please specify a folder")
+    process.exit(1)
+}
+
+console.log('Args: ', params.values)
+
 const tester = new FrontierTester({
-    queryType: "chain",
+    queryType: params.values.query!,
     // nWorkers: 20,
-    replicationFactor: 3,
-    chainLength: 1,
-    numTuples: 20000,
-    warmUpTuples: 0,
+    replicationFactor: +params.values.replication!,
+    chainLength: +params.values.length!,
+    numTuples:  +params.values.tuples!,
+    warmUpTuples: +params.values.warmup!,
+    folder: params.values.folder!,
+    keepalive: params.values.keepalive ?? false,
 })
 tester.run()
